Best animation system

+++ src/main/java/fr/theshark34/swinger/animation/QueryLoopAction.java
@@ -2,6 +2,6 @@ package fr.theshark34.swinger.animation;
 
 public interface QueryLoopAction {
 
-    void onLoop(int query);
+    void onLoop(long query);
 
 }
